+++ R/zzz.R
@@ -1,6 +1,7 @@
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage(
-    "vaccine (version 0.1.0).\nType ?vaccine to get started."
-  )
+  vrs <- utils::packageVersion("vaccine")
+  packageStartupMessage(paste0(
+    "vaccine (version ", vrs, ").\nType ?vaccine to get started."
+  ))
 }