+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ### Minor changes
 
+- Fixed a bug that displayed an incorrect package version number at startup.
 - Fixed a bug related to `plot_ce(..., density_type="kde edge")`.
 - Fixed a minor bug related to the `SuperLearner` package introduced in version 1.2.0.